Added validation layers and testing the graphics engine if everything is working fine.
Added the validation layers, and they will be packed into the final APK only on the debug mode. The graphics engine has a tiny issue with the queues, because of it it's not added to this commit (its commented out), but everything else, the UI components are working fine. Added a new camera to support dart based camera manipulation for VR, but its gonna get extended to support directly the backend camera, which is expected to perform better than Dart, as its written in C++ and uses SIMD.