Updated the graphics engine and broke the libs directory into two to better debug the application.
The Firefly API was changed, and was improved and the graphics engine was updated accordingly to support the new API. The library files which will be included in the APK package is broken into two, on containing everything and another containing just the clang shared library. This is because now we can indipendently debug the graphics engine and the Vulkan engine to see if we have any other undefined behaviours in our application.