Optimized and fixed a minor issue with the graphics instance creation. Fixed a CI/CD workflow issue.
Optimized the shader code loading form runtime loading to compile time. The graphics instance creation error was also fixed, basically we had to set the validation to false where as previously we were not properly setting it. The github workflow was also fixed, previously we set the branch as master, but we only had a branch named main. This issue is now resolved.